+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - keep last `drivetrain/hvBatteryActive` state until it's updated from the API
   - allow setting the `drivetrain/hvBatteryActive/set` state to force updates
   - forbid retained set messages
+  - added topics `refresh/lastVehicleState` and `refresh/lastChargeState`
 
 ## [0.1.0] - 2023-03-29
 ### Added